pyproject.toml
[project]
name = "llm-model-zoo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=6.5.1",
    "huggingface-hub>=1.4.1",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "safetensors>=0.7.0",
    "transformers>=5.1.0",
]
